fixed issue with x and y coordinate types

+++ frontend/src/components/JSONEntrySidebarContent.js
@@ -27,8 +27,8 @@ const nodeSchema = {
             }
         }},
         "resourceRules": {"type": "array"},
-        "x": { "type": "integer"},
-        "y":{ "type": "integer"}
+        "x": { "type": "float"},
+        "y":{ "type": "float"}
     }
    };
